pyproject.toml
[tool.poetry]
name = "cowin-notify"
version = "1.0.0"
description = "dddd"
authors = ["Mohammed Rishad <mohammedrishad.vk@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
requests = "^2.25.1"
python-environ = "^0.4.54"
Jinja2 = "^2.11.3"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
